created utils and moved compression and formatting there
